update cmake optional install directory

+++ trunk/src/GUI/playlist/view/PlaylistView.cpp
@@ -349,7 +349,7 @@ void PlaylistView::fill(MetaDataList &v_metadata, int cur_play_idx){
 
         QModelIndex model_idx = _model->index(i, 0);
 
-        md.pl_playing = (cur_play_idx == i);
+        md.pl_playing = (cur_play_idx == (int) i);
         if(md.pl_playing) idx_cur_playing = model_idx;
 
         if(md.pl_selected)
